update error ui messages for software package and edit packages
