--- frontend/pages/SoftwarePage/SoftwareTitleDetailsPage/EditSoftwareModal/helpers.ts
@@ -1,13 +1,0 @@
-import { getErrorReason } from "interfaces/errors";
-
-const UPLOAD_ERROR_MESSAGES = {
-  default: {
-    message: "Couldn't edit software. Please try again.",
-  },
-};
-
-// eslint-disable-next-line import/prefer-default-export
-export const getErrorMessage = (err: unknown) => {
-  if (typeof err === "string") return err;
-  return getErrorReason(err) || UPLOAD_ERROR_MESSAGES.default.message;
-};
